requirements.txt
# Core ML and Data Processing
torch
torchvision
torchaudio
torchmetrics>=1.0.0
lightning>=2.0.0
scikit-learn>=1.3.0
# numpy==1.26.4
scipy
pandas>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
face-recognition>=1.3.0
dlib>=19.24.0

# Deep Learning for Emotion Recognition
transformers>=4.30.0
datasets>=2.14.0

# AI/LLM Integration
ollama>=0.1.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# VR/AR Development
pygame>=2.5.0
pyglet>=1.5.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0
wandb>=0.15.0

# Audio Analysis
pydub>=0.25.1
webrtcvad>=2.0.10

# Web Application
streamlit